+++ Source/Core/Core/Debugger/Debugger_SymbolMap.cpp
@@ -12,7 +12,7 @@
 #include "Common/StringUtil.h"
 
 #include "Core/Core.h"
-#include "Core/HW/Memmap.h"
+#include "Core/PowerPC/MMU.h"
 #include "Core/PowerPC/PPCAnalyst.h"
 #include "Core/PowerPC/PPCSymbolDB.h"
 #include "Core/PowerPC/PowerPC.h"
