+++ Source/Core/Core/PowerPC/Interpreter/Interpreter_LoadStore.cpp
@@ -12,6 +12,7 @@
 #include "Core/PowerPC/Interpreter/Interpreter.h"
 #include "Core/PowerPC/Interpreter/Interpreter_FPUtils.h"
 #include "Core/PowerPC/JitInterface.h"
+#include "Core/PowerPC/MMU.h"
 #include "Core/PowerPC/PowerPC.h"
 
 bool Interpreter::m_reserve;
