+++ Source/Core/Core/PowerPC/Interpreter/Interpreter_LoadStorePaired.cpp
@@ -12,6 +12,7 @@
 #include "Common/MathUtil.h"
 #include "Core/PowerPC/Interpreter/Interpreter.h"
 #include "Core/PowerPC/Interpreter/Interpreter_FPUtils.h"
+#include "Core/PowerPC/MMU.h"
 #include "Core/PowerPC/PowerPC.h"
 
 // dequantize table
