+++ Source/Core/Core/PowerPC/SignatureDB/MEGASignatureDB.cpp
@@ -16,8 +16,8 @@
 #include "Common/Logging/Log.h"
 #include "Common/StringUtil.h"
 
+#include "Core/PowerPC/MMU.h"
 #include "Core/PowerPC/PPCSymbolDB.h"
-#include "Core/PowerPC/PowerPC.h"
 
 namespace
 {
