+++ Source/Core/DolphinQt2/Debugger/CodeViewWidget.cpp
@@ -22,6 +22,7 @@
 #include "Common/StringUtil.h"
 #include "Core/Core.h"
 #include "Core/Debugger/PPCDebugInterface.h"
+#include "Core/PowerPC/MMU.h"
 #include "Core/PowerPC/PPCAnalyst.h"
 #include "Core/PowerPC/PPCSymbolDB.h"
 #include "Core/PowerPC/PowerPC.h"
