+++ Source/Core/DolphinQt2/Debugger/CodeWidget.cpp
@@ -19,6 +19,7 @@
 #include "Core/Core.h"
 #include "Core/Debugger/Debugger_SymbolMap.h"
 #include "Core/HW/CPU.h"
+#include "Core/PowerPC/MMU.h"
 #include "Core/PowerPC/PPCSymbolDB.h"
 #include "Core/PowerPC/PowerPC.h"
 #include "DolphinQt2/Host.h"
