+++ Source/Core/DolphinQt2/MenuBar.cpp
@@ -28,12 +28,12 @@
 #include "Core/HLE/HLE.h"
 #include "Core/HW/WiiSave.h"
 #include "Core/HW/Wiimote.h"
-#include "Core/Host.h"
 #include "Core/IOS/ES/ES.h"
 #include "Core/IOS/IOS.h"
 #include "Core/IOS/USB/Bluetooth/BTEmu.h"
 #include "Core/Movie.h"
 #include "Core/PowerPC/JitInterface.h"
+#include "Core/PowerPC/MMU.h"
 #include "Core/PowerPC/PPCAnalyst.h"
 #include "Core/PowerPC/PPCSymbolDB.h"
 #include "Core/PowerPC/PowerPC.h"
