+++ Source/Core/DolphinWX/Debugger/MemoryView.cpp
@@ -24,8 +24,8 @@
 #include "Common/StringUtil.h"
 #include "Core/Core.h"
 #include "Core/HW/Memmap.h"
+#include "Core/PowerPC/MMU.h"
 #include "Core/PowerPC/PPCSymbolDB.h"
-#include "Core/PowerPC/PowerPC.h"
 #include "DolphinWX/Debugger/CodeWindow.h"
 #include "DolphinWX/Debugger/DebuggerUIUtil.h"
 #include "DolphinWX/Debugger/WatchWindow.h"
